# api-doc/swagger.yaml
basePath: /api
...
host: pmp-dev.alki.earth
info:
  contact: {}
  description: API server to manage Backpack Inventory and Packing Lists
  title: PimpMyPack API
  version: "1.0"
paths:
...
  /v1/myinventory/{id}/image:
    delete:
      description: Delete the image for an inventory item. Only the item owner can
        delete images.
      parameters:
      - description: Inventory Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Image deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid item ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Item does not belong to user
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Item not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete inventory item image
      tags:
      - Inventory Images
    get:
      description: Get the image for an inventory item. Only the item owner can view
        the image.
      parameters:
      - description: Inventory Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Invalid item ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Item not found or has no image
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get inventory item image
      tags:
      - Inventory Images
    post:
      consumes:
      - multipart/form-data
      description: Upload or update an image for an inventory item. Only the item
        owner can upload images.
      parameters:
      - description: Inventory Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file (JPEG, PNG, or WebP, max 5MB)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image uploaded successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or image
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Item does not belong to user
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Item not found
          schema:
            additionalProperties:
              type: string
            type: object
        "413":
          description: Payload too large
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Upload or update inventory item image
      tags:
      - Inventory Images
...
schemes:
- https
swagger: "2.0"